+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Master](https://github.com//AppwiseCore)
 
+### Bug Fixes
+
+* Core: Fix crash in CocoaLumberjack Crashlytics logger.
+
 ## [0.6.0](https://github.com/appwise-labs/AppwiseCore/releases/tag/0.6.0)
 
 ### New Features